+++ lib/src/loading/_loading.dart
@@ -1,0 +1,8 @@
+// Created by Muhamad Fauzi Ridwan on 07/11/21.
+
+import 'package:flutter/material.dart';
+import 'package:flutter_spinkit/flutter_spinkit.dart';
+
+part 'circle_loading.dart';
+part 'linear_loading.dart';
+part 'spinkit_loading.dart';