+++ lib/src/view/waiting_view.dart
@@ -1,0 +1,13 @@
+// Created by Muhamad Fauzi Ridwan on 07/11/21.
+
+part of '_view.dart';
+
+/// a waiting view for waiting state on states rebuilder.
+class WaitingView extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return ListView(
+      children: [],
+    );
+  }
+}